+++ src/main/java/com/quizz/pro/dao/UserDAOImpl.java
@@ -1,4 +1,4 @@
-package com.quizz.pro.DAO;
+package com.quizz.pro.dao;
 
 import java.util.List;
 
@@ -14,7 +14,7 @@ import org.springframework.orm.hibernate5.HibernateTemplate;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.quizz.pro.Entity.User;
+import com.quizz.pro.entity.User;
 
 @Repository
 @Profile("dev")
