+++ src/main/java/com/quizz/pro/util/WebConfig.java
@@ -1,4 +1,4 @@
-package com.quizz.pro.Util;
+package com.quizz.pro.util;
 
 import java.util.Properties;
